new code 7 nov
git commit fourth

+++ src/controllers/ClientController.ts
@@ -1,32 +1,37 @@
 import {Client} from "../entities/Client";
+import {getClientRepository} from "../Repository/Repository";
 import {IRouterContext} from "koa-router";
 import {
     BAD_REQUEST_MESSAGE,
     BAD_REQUEST_STATUS, CREATED_STATUS, CREATED_STATUS_MESSAGE,
-    INTERNAL_SERVER_ERROR_MESSAGE,
-    INTERNAL_SERVER_ERROR_STATUS, NOT_FOUND_MESSAGE, NOT_FOUND_STATUS, OK_STATUS, OK_STATUS_MESSAGE
-} from "../utils/statusCode";
-import {BaseController} from "./baseController";
+    NOT_FOUND_MESSAGE, NOT_FOUND_STATUS, OK_STATUS, OK_STATUS_MESSAGE
+} from "../utils/StatusCode";
+import {BaseController} from "./BaseController";
+import {Repository} from "typeorm";
 
 export class ClientController extends BaseController {
-    protected clientDataRepo: any;
+    protected clientDataRepo: Repository<Client>;
 
-    constructor(connection: any) {
+    constructor() {
         super();
-        this.clientDataRepo = connection.getRepository(Client);
-
-
+        (async () => {
+            this.clientDataRepo = await getClientRepository();
+        })();
     }
 
     //Create new client
     public async createClient(ctx: any) {
 
         const {name, email} = ctx.request.body as { name: string; email: string };
+
+        if (!name && !email) {
+            return this.badRequest(ctx, BAD_REQUEST_STATUS, BAD_REQUEST_MESSAGE);
+        }
+
         const client = new Client();
         client.name = name;
         client.email = email;
-        console.log(this.clientDataRepo);
-        console.log(await this.clientDataRepo.save(client));
+        await this.clientDataRepo.save(client);
         this.okStatus(ctx, CREATED_STATUS, CREATED_STATUS_MESSAGE);
     }
 
@@ -38,9 +43,11 @@ export class ClientController extends BaseController {
                 products: true
             }
         });
-        if(!clients){
+
+        if (!clients) {
             this.badRequest(ctx, BAD_REQUEST_STATUS, BAD_REQUEST_MESSAGE)
         }
+
         this.okStatus(ctx, OK_STATUS, OK_STATUS_MESSAGE)
         ctx.body = clients;
     }
@@ -49,6 +56,11 @@ export class ClientController extends BaseController {
     //Get Clients by id
     public async getClientById(ctx: IRouterContext) {
         const id = +ctx.params.id;
+
+        if (isNaN(+id) || +id <= 0) {
+            return this.badRequest(ctx, BAD_REQUEST_STATUS, BAD_REQUEST_MESSAGE);
+        }
+
         const clients = await this.clientDataRepo.findOne({
             where: {
                 id: id
@@ -57,10 +69,12 @@ export class ClientController extends BaseController {
                 products: true
             }
         });
+
         if (!clients) {
             this.badRequest(ctx, NOT_FOUND_STATUS, NOT_FOUND_MESSAGE)
             return;
         }
+
         ctx.body = clients;
         this.okStatus(ctx, OK_STATUS, OK_STATUS_MESSAGE)
     }
@@ -69,16 +83,23 @@ export class ClientController extends BaseController {
     //Update Client by id
     public async updateClientById(ctx: IRouterContext) {
         const id = +ctx.params.id;
+
+        if (isNaN(id) || id <= 0 || !id) {
+            return this.badRequest(ctx, BAD_REQUEST_STATUS, BAD_REQUEST_MESSAGE);
+        }
+
         const {name, email} = ctx.request.body as { name: string; email: string };
         const client = await this.clientDataRepo.findOne({
             where: {
                 id: id
             }
         })
+
         if (!client) {
             this.badRequest(ctx, BAD_REQUEST_STATUS, BAD_REQUEST_MESSAGE)
             return;
         }
+
         await this.clientDataRepo.update(id, {
             name: name,
             email: email
@@ -90,6 +111,21 @@ export class ClientController extends BaseController {
     //Delete Client by id
     public async deleteClientById(ctx: IRouterContext) {
         const id = +ctx.params.id;
+
+        if (isNaN(id) || id <= 0 || (!id)) {
+            return this.badRequest(ctx, BAD_REQUEST_STATUS, BAD_REQUEST_MESSAGE);
+        }
+
+        const clientData = await this.clientDataRepo.findOne({
+            where: {
+                id: id
+            }
+        })
+
+        if(!clientData){
+            return this.badRequest(ctx, NOT_FOUND_STATUS, NOT_FOUND_MESSAGE);
+        }
+
         const deletedData = await this.clientDataRepo.delete({id});
         if (deletedData.affected === 0) {
             this.badRequest(ctx, NOT_FOUND_STATUS, NOT_FOUND_MESSAGE)
