+++ src/controllers/ProductController.ts
@@ -1,39 +1,50 @@
 import { Client } from "../entities/Client";
-import { DBConnection } from "../db/dbConnection";
 import { IRouterContext } from "koa-router";
 import { Product } from "../entities/Product";
-import { Options } from "../config/type";
+import { Options } from "../config/Type";
 import {
     BAD_REQUEST_MESSAGE,
     BAD_REQUEST_STATUS,
     INTERNAL_SERVER_ERROR_MESSAGE,
     INTERNAL_SERVER_ERROR_STATUS, NOT_FOUND_MESSAGE, NOT_FOUND_STATUS, OK_STATUS, OK_STATUS_MESSAGE
-} from "../utils/statusCode";
-import {BaseController} from "./baseController";
-
-const dbConnect = new DBConnection();
+} from "../utils/StatusCode";
+import {BaseController} from "./BaseController";
+import {getClientRepository, getProductRepository} from "../Repository/Repository";
+import {Repository} from "typeorm";
 
 export class ProductController extends BaseController{
-    protected productDataRepo : any;
-    constructor(connection: any) {
+    protected productDataRepo : Repository<Product>;
+    protected clientDataRepo : Repository<Client>;
+    constructor() {
         super();
-        this.productDataRepo = connection.getRepository(Product);
+        (async () => {
+            this.productDataRepo = await getProductRepository();
+            this.clientDataRepo = await getClientRepository()
+        })();
     }
 
     // Create Product for a Client
     public async createProductForClient(ctx: IRouterContext) {
-        const clientDataRepo = (await dbConnect.connect()).getRepository(Client);
-
         const id = ctx.params.id;
         const { name, description, options } = ctx.request.body as { name: string; description: string; options?: Options };
 
-        // Validation: Ensure description is provided
-        if (!description || description.trim() === "") {
-            this.badRequest(ctx, BAD_REQUEST_STATUS, BAD_REQUEST_MESSAGE);
-            return;
+        if (isNaN(+id) || +id <= 0) {
+            return this.badRequest(ctx, BAD_REQUEST_STATUS, BAD_REQUEST_MESSAGE);
         }
 
-        const client = await clientDataRepo.findOneOrFail({
+        if (!description || description.trim() === "") {
+            return this.badRequest(ctx, BAD_REQUEST_STATUS, BAD_REQUEST_MESSAGE);
+        }
+
+        if(!name || name.trim() === "") {
+            return this.badRequest(ctx, BAD_REQUEST_STATUS, BAD_REQUEST_MESSAGE);
+        }
+
+        if(!options){
+            return this.badRequest(ctx, BAD_REQUEST_STATUS, BAD_REQUEST_MESSAGE);
+        }
+
+        const client = await this.clientDataRepo.findOneOrFail({
             where: { id: +id }
         });
 
@@ -54,6 +65,11 @@ export class ProductController extends BaseController{
     // Get All Products for a Client
     public async getProductForClient(ctx: IRouterContext) {
         const clientId = ctx.params.id;
+
+        if(!clientId){
+            return this.badRequest(ctx, BAD_REQUEST_STATUS, BAD_REQUEST_MESSAGE)
+        }
+
         try {
             const products = await this.productDataRepo.findOneOrFail({
                 where: {
@@ -63,9 +79,9 @@ export class ProductController extends BaseController{
                     },
                 relations: { variant: true },  // Include variants if needed
             });
+
             if (!products) {
-                this.badRequest(ctx, NOT_FOUND_STATUS, NOT_FOUND_MESSAGE);
-                return;
+                return this.badRequest(ctx, NOT_FOUND_STATUS, NOT_FOUND_MESSAGE);
             }
 
             this.okStatus(ctx, OK_STATUS, OK_STATUS_MESSAGE);
@@ -76,8 +92,12 @@ export class ProductController extends BaseController{
     }
 
     // Get Product by Id
-    public async getProductById(ctx: IRouterContext) {
+     async getProductById(ctx: IRouterContext) {
         const id = +ctx.params.id;
+
+         if(!id){
+             return this.badRequest(ctx, BAD_REQUEST_STATUS, BAD_REQUEST_MESSAGE)
+         }
 
         try {
             const product = await this.productDataRepo.find({
@@ -86,8 +106,7 @@ export class ProductController extends BaseController{
             });
 
             if (!product || product.length === 0) {
-                this.badRequest(ctx, NOT_FOUND_STATUS, NOT_FOUND_MESSAGE);
-                return;
+                return this.badRequest(ctx, NOT_FOUND_STATUS, NOT_FOUND_MESSAGE);
             }
 
             this.okStatus(ctx, OK_STATUS, OK_STATUS_MESSAGE);
@@ -100,14 +119,18 @@ export class ProductController extends BaseController{
     // Update product by id
     public async updateProductById(ctx: IRouterContext) {
         const id = +ctx.params.id;
+
+        if(isNaN(id) || id <= 0){
+            return this.badRequest(ctx, BAD_REQUEST_STATUS, BAD_REQUEST_MESSAGE)
+        }
+
         const { name, description } = ctx.request.body as { name: string; description: string };
 
         try {
             const product = await this.productDataRepo.findOneBy({ id });
 
             if (!product) {
-                this.badRequest(ctx, NOT_FOUND_STATUS, NOT_FOUND_MESSAGE)
-                return;
+                return this.badRequest(ctx, NOT_FOUND_STATUS, NOT_FOUND_MESSAGE)
             }
 
             // Update fields
@@ -127,12 +150,15 @@ export class ProductController extends BaseController{
     public async deleteProductById(ctx: IRouterContext) {
         const id = +ctx.params.id;
 
+        if(isNaN(id) || id <= 0){
+            return this.badRequest(ctx, BAD_REQUEST_STATUS, BAD_REQUEST_MESSAGE)
+        }
+
         try {
             const result = await this.productDataRepo.delete(id);
 
             if (result.affected === 0) {
-                this.badRequest(ctx, NOT_FOUND_STATUS, NOT_FOUND_MESSAGE);
-                return;
+                return this.badRequest(ctx, NOT_FOUND_STATUS, NOT_FOUND_MESSAGE);
             }
 
             this.okStatus(ctx, OK_STATUS, OK_STATUS_MESSAGE)
